feat: add support for Gradle 8.14.5 feat: add support for Kotlin 2.3 feat: add support for shadowJar 9.2
# Conflicts:
#	build.gradle
#	src/main/java/me/eldodebug/soar/discord/DiscordRPC.java
#	src/main/java/me/eldodebug/soar/gui/gamemenus/MenuManager.java
#	src/main/java/me/eldodebug/soar/gui/mainmenu/GuiGlideMainMenu.java
#	src/main/java/me/eldodebug/soar/injection/mixin/GlideTweaker.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/ArrayListMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/CalendarMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/CoordsMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/HorseStatsMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/InternalSettingsMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/InventoryDisplayMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/Items2DMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/KillEffectsMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/MemoryUsageMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/MouseStrokesMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/PackDisplayMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/PlayerListMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/PotionStatusMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/RearviewMod.java
#	src/main/java/me/eldodebug/soar/management/mods/impl/SoundSubtitlesMod.java

+++ src/main/java/me/eldodebug/soar/management/mods/impl/ClockMod.java
@@ -12,16 +12,14 @@ import me.eldodebug.soar.management.nanovg.font.Icons;
 import org.lwjgl.nanovg.NVGColor;
 import org.lwjgl.nanovg.NanoVG;
 
-import me.eldodebug.soar.Glide;
+import com.glideclient.Glide;
 import me.eldodebug.soar.management.event.EventTarget;
-import me.eldodebug.soar.management.event.impl.EventRender2D;
 import me.eldodebug.soar.management.language.TranslateText;
 import me.eldodebug.soar.management.mods.SimpleHUDMod;
 import me.eldodebug.soar.management.mods.settings.impl.BooleanSetting;
 import me.eldodebug.soar.management.mods.settings.impl.ComboSetting;
 import me.eldodebug.soar.management.mods.settings.impl.combo.Option;
 import me.eldodebug.soar.management.nanovg.NanoVGManager;
-import me.eldodebug.soar.management.nanovg.font.Fonts;
 import me.eldodebug.soar.management.nanovg.font.LegacyIcon;
 
 public class ClockMod extends SimpleHUDMod {
@@ -63,7 +61,7 @@ public class ClockMod extends SimpleHUDMod {
 		
 		float scale = this.getScale();
 		int size = 128;
-		float center = size / 2;
+		float center = (float) size / 2;
 		float scaledSize = 128 * scale;
 		float radius = center * this.getScale();
 		String[] numbers = {"3", "6", "9", "12"};
